Revert "ENH/TST: Add BaseInterfaceTests tests for ArrowExtensionArray (#47377)"
This reverts commit d40c37192b3d38a3c9a0d248dc612ac9dd56e506.
