+++ pandas/tests/extension/arrow/arrays.py
@@ -23,6 +23,7 @@ from pandas.api.extensions import (
     take,
 )
 from pandas.api.types import is_scalar
+from pandas.core.arraylike import OpsMixin
 from pandas.core.arrays.arrow import ArrowExtensionArray as _ArrowExtensionArray
 from pandas.core.construction import extract_array
 
@@ -71,7 +72,7 @@ class ArrowStringDtype(ExtensionDtype):
         return ArrowStringArray
 
 
-class ArrowExtensionArray(_ArrowExtensionArray):
+class ArrowExtensionArray(OpsMixin, _ArrowExtensionArray):
     _data: pa.ChunkedArray
 
     @classmethod
